requirements.txt
# requirements.txt
python-telegram-bot==20.0
boto3==1.26.0
python-dotenv==0.21.0
